Added JWT Authentication and wrapped up backend

+++ backend/controllers/goalControllers.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../model/goalsModel");
+const User = require("../model/userModel");
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -12,6 +13,7 @@ const setGoal = asyncHandler(async (req, res) => {
   } else {
     const goal = await Goal.create({
       text: req.body.text,
+      user: req.user.id,
     });
     res.status(200).json(goal);
   }
@@ -23,8 +25,18 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
-  } 
-  else {
+  } else {
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+      res.status(401);
+      throw new Error("User not found");
+    }
+
+    if (goal.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error("Unauthorized User");
+    }
     var updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
     });
@@ -35,15 +47,28 @@ const updateGoal = asyncHandler(async (req, res) => {
 
 const deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
+  console.log(req.params);
+  console.log(goal);
 
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
   }
-  
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Unauthorized User");
+  }
+
   await goal.deleteOne();
 
-  res.status(200).json({ id: req.params.id});
+  res.status(200).json({ id: req.params.id });
 });
 
 module.exports = { getGoals, setGoal, updateGoal, deleteGoal };
